Move common functionality into Rotatable.  Remove complex properties and propertyinstance classes.
